pyproject.toml
[project]
name = "financeapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "postgres>=4.0",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
